appRefactoring
appRefactoring, grouping to folders
