+++ Tandauly/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Contents.json
