+++ Tandauly/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Icon-App-20x20@1x.png
