+++ Tandauly/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Icon-App-20x20@2x.png
