+++ Tandauly/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Icon-App-40x40@3x.png
