+++ Tandauly/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Icon-App-60x60@2x.png
