+++ Tandauly/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Icon-App-76x76@2x.png
