+++ Tandauly/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Icon-App-83.5x83.5@2x.png
