+++ Tandauly/SupportingFiles/Assets.xcassets/Contents.json
