+++ Tandauly/SupportingFiles/Assets.xcassets/can.imageset/Contents.json
