+++ Tandauly/SupportingFiles/Assets.xcassets/mapImage.imageset/map@2x.png
