+++ Tandauly/SupportingFiles/Assets.xcassets/mapImage.imageset/map@3x.png
