+++ Tandauly/SupportingFiles/Assets.xcassets/navigator.imageset/navigator.png
