+++ Tandauly/SupportingFiles/Assets.xcassets/pin.imageset/pin.png
