+++ Tandauly/SupportingFiles/Assets.xcassets/star.imageset/Contents.json
