+++ Tandauly/SupportingFiles/Assets.xcassets/star1.imageset/Contents.json
